optimization #2:  in `modules/load_data.py > load_sample` for the np array `target_full_array` which has only 0s and 1s set its type to np.int8 instead of np.float64/np.int64 (default), which saves memory  reduced peak mem usage by 4.2%

+++ modules/load_data.py
@@ -184,7 +184,7 @@ def load_sample(
 
     """
     sample_index = get_sample_index(sample, samples_txt_path=f"./data/SI_data/samples.txt")
-    target_full_array = np.zeros((chr_length,2))
+    target_full_array = np.zeros((chr_length, 2), dtype=np.int8)
     target_full_array[:,0] = np.unpackbits(ref_panel_full_array_full_packed[:,sample_index[0]])[:chr_length]
     target_full_array[:,1] = np.unpackbits(ref_panel_full_array_full_packed[:,sample_index[1]])[:chr_length]
